+++ R/get_ontology_github.R
@@ -11,27 +11,56 @@ get_ontology_github <- function(name,
                                 filetype=".obo",
                                 file=paste0(name,filetype),  
                                 tag="latest",
-                                save_dir=cache_dir(),
+                                
+                                add_metadata=TRUE,
+                                add_n_edges=TRUE,
+                                add_ontology_levels=TRUE,
+                                lvl=2,
+                                
+                                save_dir=cache_dir(subdir = "ontologies"),
                                 force_new=FALSE,
                                 ...){
   
-  save_path <- file.path(save_dir,file)
-  if(!file.exists(save_path) || isTRUE(force_new)){
-    #### From Releases #### 
+  messager("Importing ontology via GitHub.")
+  #### Get latest release/tag/version ####  
+  if(is.null(tag) || tag == "latest"){
     requireNamespace("piggyback") 
-    if(is.null(tag)){
-      messager("Identifying latest release for:",repo)
-      releases <- piggyback::pb_releases(repo = repo)
-      tag <- releases$release_name[1]
-    } 
+    messager("Identifying latest release for:",repo)
+    releases <- piggyback::pb_releases(repo = repo)
+    tag <- releases$tag_name[1]
+  } 
+  
+  save_path <- .get_ontology_savepath(save_dir=save_dir, 
+                                      path=file, 
+                                      version=tag)
+  
+  if(!file.exists(save_path) || isTRUE(force_new)){
+    requireNamespace("piggyback") 
     URL <- piggyback::pb_download_url(file = file, 
                                       repo = repo,
                                       tag = tag)
+    
     messager("Preparing ontology_index object from:",URL)
     ont <- simona::import_ontology(file=URL,
                                    ...) 
-    saveRDS(ont,save_path)
+    
+    #### Add metadata #### 
+    if(isTRUE(add_metadata)) {
+      ont <- add_ontology_metadata(ont,
+                                   add_n_edges=add_n_edges,
+                                   lvl=lvl, 
+                                   add_ontology_levels=add_ontology_levels)
+    }
+    
+    #### Cache file ####
+    if(!is.null(save_dir)){
+      dir.create(save_dir, recursive=TRUE, showWarnings = FALSE)
+      messager("Caching file -->",save_path)
+      saveRDS(ont,save_path)
+    }
+    
   } else {
+    #### Import cached file ####
     messager("Importing cached file:",save_path)
     ont <- readRDS(save_path)
   }