+++ R/plot_upheno_scatterplot.R
@@ -2,6 +2,8 @@ plot_upheno_scatterplot <- function(plot_dat){
 
   requireNamespace("ggplot2")
   n_genes_intersect <- n_genes_db1 <- equivalence_score <- NULL;
+  
+  messager("Creating UPHENO scatterplot")
   #### Check if there's a relationship between phenotype mapping scores
   # and number of shared genes ####
   ggplot2::ggplot(plot_dat,